requirements.txt
bigfloat==0.3.0
certifi==2017.4.17
chardet==3.0.4
click==6.7
cycler==0.10.0
graphviz==0.7.1
h5py==2.7.0
idna==2.5
itsdangerous==0.24
Jinja2==2.9.6
Keras==2.0.4
MarkupSafe==1.0
matplotlib==2.0.2
numpy==1.13.0
olefile==0.44
Pillow==6.2.0
protobuf==3.3.0
pydot==1.2.3
pydot-ng==1.0.0
pyparsing==2.2.0
python-dateutil==2.6.0
pytz==2017.2
requests==2.20.0
scipy==0.19.0
simplejson==3.10.0
six==1.10.0
tensorflow==1.12.2
Theano==0.9.0
tinydb==3.2.3
urllib3==1.24.2
Werkzeug==0.15.3
